feat: Utility Function GetGoogleSheetData added which reads responses of Google Form from Google Sheet
feat: Utility Function GetGoogleSheetData added which reads responses of Google Form from Google Sheet

GetGoogleSheetData() method requires the name of the Google Sheet which is mapped behind any Google Form to collect its responses. In addition to it, the method requires the number of columns and the maximim row ids already present.

For Online Data Entry, Google Forms is very easy to setup with all responses been collected into a Google Spreadsheet. This enables us to get the data entry process started, however, when multiple data entry forms are designed and there is a need of further data processing or action against the new record been entered, we are limited to work on Spreadsheets only. This utility helps to read Google Spreadsheet into a dataset and persist it into an XML file for records. Additionally, we could use this data set to persist data into respective table on any local SQL Server.

Below assumptions are kept into consideration:

1) Generally, new fields are not added to data entry forms so frequently, hence mentioning the column count is considered as setup part.

2) Max Rows is an incremental value and it is assumed that when this utility is used in live applications this value is maintained as setup information and updated whenever new rows are read.

3) Overall, objective is to simply sync the Google Sheet data to a DataSet so that it could be further used for any business logic.